+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Add `port` argument to Cloud Run deployments via `cr_buildstep_run()`
 * Add `cr_deploy_pkgdown` and `cr_deploy_packagetests` add subsequent buildsteps to aid R package development.
+* Fix `cr_buildstep_r()` so it can run R scripts from a filename when `r_source="runtime"` (#45 - thanks @j450h1 and @axel-analyst)
 
 # googleCloudRunner 0.1.1